pyproject.toml
[tool.poetry]
name = "minesweeper"
version = "0.1.0"
description = ""
authors = ["Hemppa <hemppa18@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pygame = "^2.3.0"
invoke = "^2.0.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
coverage = "^7.2.2"
pylint = "^2.17.2"
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
